+++ runtime/config.go
@@ -10,19 +10,16 @@ type Config struct {
 	AWSSecretAccessKey string `help:"secret access key to use for AWS services"`
 	AWSRegion          string `help:"region to use for AWS services, e.g. us-east-1"`
 
-	S3Endpoint string `help:"S3 endpoint we will write archives to"`
-	S3Bucket   string `help:"S3 bucket we will write archives to"`
-	S3Minio    bool   `help:"S3 is actually Minio or other compatible service"`
+	S3Endpoint  string `help:"S3 endpoint we will write archives to"`
+	S3Bucket    string `help:"S3 bucket we will write archives to"`
+	S3PathStyle bool   `help:"S3 should use path style URLs"`
 
 	TempDir       string `help:"directory where temporary archive files are written"`
-	KeepFiles     bool   `help:"whether we should keep local archive files after upload (default false)"`
-	UploadToS3    bool   `help:"whether we should upload archive to S3"`
 	CheckS3Hashes bool   `help:"whether to check S3 hashes of uploaded archives before deleting records"`
 
 	ArchiveMessages bool   `help:"whether we should archive messages"`
 	ArchiveRuns     bool   `help:"whether we should archive runs"`
 	RetentionPeriod int    `help:"the number of days to keep before archiving"`
-	Delete          bool   `help:"whether to delete messages and runs from the db after archival (default false)"`
 	StartTime       string `help:"what time archive jobs should run in UTC HH:MM "`
 	Once            bool   `help:"whether archiver should run once and exit (default false)"`
 
@@ -43,19 +40,16 @@ func NewDefaultConfig() *Config {
 		AWSSecretAccessKey: "",
 		AWSRegion:          "us-east-1",
 
-		S3Endpoint: "https://s3.amazonaws.com",
-		S3Bucket:   "temba-archives",
-		S3Minio:    false,
+		S3Endpoint:  "https://s3.amazonaws.com",
+		S3Bucket:    "temba-archives",
+		S3PathStyle: false,
 
 		TempDir:       "/tmp",
-		KeepFiles:     false,
-		UploadToS3:    true,
 		CheckS3Hashes: true,
 
 		ArchiveMessages: true,
 		ArchiveRuns:     true,
 		RetentionPeriod: 90,
-		Delete:          false,
 		StartTime:       "00:01",
 		Once:            false,